Rearrange files and namespaces for sanity
