+++ lib/markovian/tools.rb
@@ -1,6 +1,7 @@
 require 'markovian'
 require "markovian/tools/version"
-require "markovian/tools/importers/twitter/csv_importer"
+require "markovian/importers/twitter/csv_importer"
+require "markovian/publishers/tweeter"
 
 module Markovian
   module Tools
